AI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
firebase-admin
tensorflow
pandas
scikit-learn
googletrans==4.0.0-rc1
azure-functions
transformers
pyparsing
latest
tf-keras